requirements.txt
nonebot-adapter-onebot
nonebot-adapter-discord
nonebot-adapter-telegram
nonebot-plugin-apscheduler
httpx
ping3
yaml